Module runtime::builtins

Builtin functions. These are usually specially generated in the parse stage, and compiled into special op codes. These operations are hard coded for the builtin types.

Functions

add

and

cmp_eq

cmp_geq

cmp_gt

cmp_leq

cmp_lt

cmp_neq

div

index_get

index_set

mod_

mul

negate

not

or

sub