Module runtime::bytecode

get compiled into to be run.

Structs

DebugSymbol

FRAME_ID_GEN

Frame

FrameIdGen

GlobalContext

Enums

Op

An operation on the technetium virtual machine

Functions

gen_frame_id

Generate a new unique FrameID. This uses a static counter behind a mutex to guarantee a unique ID.

Type Definitions

ContextId

Unique to each function in source code. Unlike FrameId, multiple frames can share a ContextID if they came from the same function of source code.

DebugSpanDescriptor

FrameId

An identifier unique to each frame of computation

GlobalConstantDescriptor

LocalName

NonLocalName

NonLocalUnmappedName