Enum runtime::error::RuntimeErrorType

pub enum RuntimeErrorType {
    TypeError,
    IntegerTooBigError,
    AttributeError,
    InternalError,
    IndexOutOfBounds,
    ChildProcessError,
    VariableUndefinedError,
    IOError,
    SysInfoError,
    PoisonError,
    BorrowError,
    BorrowMutError,
    MutateImmutableError,
    KeyError,
    OpenError,
}

Variants

TypeError

IntegerTooBigError

AttributeError

InternalError

IndexOutOfBounds

ChildProcessError

VariableUndefinedError

Caused by trying to read an uninitialized variable

IOError

Wrapping an std::io::Error

SysInfoError

An error raised by the sys_info crate

PoisonError

An error raised by trying to lock() a poisoned mutex on an Object

BorrowError

An error raised by trying to modify and read something at the same time

BorrowMutError

MutateImmutableError

KeyError

Caused by trying to read a key that doesn't exist in a dictionary

OpenError