Module runtime::metadata

Modules

asset_registry

aura

aura_ext

authorship

base_fee

btc_relay

clients_info

collator_selection

contracts

cumulus_xcm

currency

democracy

dex_general

dex_stable

dex_swap_router

dmp_queue

escrow

escrow_annuity

escrow_rewards

ethereum

evm

evm_chain_id

farming

farming_rewards

fee

identity

issue

loans

multi_transaction_payment

multisig

nomination

oracle

parachain_info

parachain_system

polkadot_xcm

preimage

proxy

redeem

replace

runtime_apis

runtime_types

scheduler

security

session

sudo

supply

system

technical_committee

technical_membership

timestamp

tokens

transaction_payment

tx_pause

unknown_tokens

utility

vault_annuity

vault_capacity

vault_registry

vault_rewards

vault_staking

vesting

x_tokens

xcmp_queue

Structs

ConstantsApi

StorageApi

TransactionApi

Enums

Call

Error

Event

Statics

PALLETS

Functions

apis

constants

storage

tx

validate_codegen

check whether the Client you are using is aligned with the statically generated codegen.

Type Definitions

DispatchError

The error type returned when there is a runtime issue.