Struct runtime::net::tcp::TcpStream

pub struct TcpStream { /* fields omitted */ }

A TCP stream between a local and a remote socket.

A TcpStream can either be created by connecting to an endpoint, via the connect method, or by accepting a connection from a listener. It can be read or written to using the AsyncRead, AsyncWrite, and related extension traits in futures::io.

The connection will be closed when the value is dropped. The reading and writing portions of the connection can also be shut down individually with the shutdown method.

Examples

#![feature(async_await, await_macro, futures_api)]

use futures::prelude::*;
use runtime::net::TcpStream;

#[runtime::main]
async fn main() -> Result<(), failure::Error> {
    let mut stream = await!(TcpStream::connect("127.0.0.1:8080"))?;
    println!("Connected to {}", &stream.peer_addr()?);

    let msg = "hello world";
    println!("<- {}", msg);
    await!(stream.write_all(msg.as_bytes()))?;

    let mut buf = vec![0u8; 1024];
    await!(stream.read(&mut buf))?;
    println!("-> {}\n", std::str::from_utf8(&mut buf)?);

    Ok(())
}

Methods

impl TcpStream

Important traits for Connect
impl Future for Connect type Output = Result<TcpStream>;

pub fn connect<A: ToSocketAddrs>(addr: A) -> Connect

Create a new TCP stream connected to the specified address.

This function will create a new TCP socket and attempt to connect it to the addr provided. The returned future will be resolved once the stream has successfully connected, or it will return an error if one occurs.

Examples

#![feature(async_await, await_macro, futures_api)]
use runtime::net::TcpStream;

let stream = await!(TcpStream::connect("127.0.0.1:0"))?;

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local address that this stream is connected to.

Examples

#![feature(async_await, await_macro, futures_api)]
use runtime::net::TcpStream;
use std::net::{IpAddr, Ipv4Addr};

let stream = await!(TcpStream::connect("127.0.0.1:8080"))?;

let expected = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));
assert_eq!(stream.local_addr()?.ip(), expected);

pub fn peer_addr(&self) -> Result<SocketAddr>

Returns the remote address that this stream is connected to.

Examples

#![feature(async_await, await_macro, futures_api)]
use runtime::net::TcpStream;
use std::net::{IpAddr, Ipv4Addr};

let stream = await!(TcpStream::connect("127.0.0.1:8080"))?;

let expected = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));
assert_eq!(stream.peer_addr()?.ip(), expected);

pub fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O on the specified portions to return immediately with an appropriate value (see the documentation of Shutdown).

Examples

#![feature(async_await, await_macro, futures_api)]

use std::net::Shutdown;
use runtime::net::TcpStream;

let stream = await!(TcpStream::connect("127.0.0.1:8080"))?;
stream.shutdown(Shutdown::Both)?;