Struct runtime::net::udp::UdpSocket

pub struct UdpSocket { /* fields omitted */ }

A UDP socket.

After creating a UdpSocket by binding it to a socket address, data can be sent to and received from any other socket address.

As stated in the User Datagram Protocol's specification in IETF RFC 768, UDP is an unordered, unreliable protocol. Refer to TcpListener and TcpStream for async TCP primitives, and std::net for synchronous networking primitives.

Examples

#![feature(async_await, await_macro, futures_api)]

use runtime::net::UdpSocket;

#[runtime::main]
async fn main() -> std::io::Result<()> {
    let mut socket = UdpSocket::bind("127.0.0.1:8080")?;
    let mut buf = vec![0u8; 1024];

    println!("Listening on {}", socket.local_addr()?);

    loop {
        let (recv, peer) = await!(socket.recv_from(&mut buf))?;
        let sent = await!(socket.send_to(&buf[..recv], &peer))?;
        println!("Sent {} out of {} bytes to {}", sent, recv, peer);
    }
}

Methods

impl UdpSocket

pub fn bind<A: ToSocketAddrs>(addr: A) -> Result<Self>

Creates a UDP socket from the given address.

Binding with a port number of 0 will request that the OS assigns a port to this socket. The port allocated can be queried via the local_addr method.

Examples

use runtime::net::UdpSocket;

let socket = UdpSocket::bind("127.0.0.1:0")?;

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local address that this listener is bound to.

This can be useful, for example, when binding to port 0 to figure out which port was actually bound.

Examples

use runtime::net::UdpSocket;

let socket = UdpSocket::bind("127.0.0.1:0")?;
println!("Address: {:?}", socket.local_addr());

Important traits for SendTo<'socket, 'buf>
impl<'socket, 'buf> Future for SendTo<'socket, 'buf> type Output = Result<usize>;

pub fn send_to<'socket, 'buf, A: ToSocketAddrs>(
&'socket mut self,
buf: &'buf [u8],
addr: A
) -> SendTo<'socket, 'buf>

Sends data on the socket to the given address.

On success, returns the number of bytes written.

Examples

#![feature(async_await, await_macro, futures_api)]
use runtime::net::UdpSocket;

const THE_MERCHANT_OF_VENICE: &[u8] = b"
    If you prick us, do we not bleed?
    If you tickle us, do we not laugh?
    If you poison us, do we not die?
    And if you wrong us, shall we not revenge?
";

let mut socket = UdpSocket::bind("127.0.0.1:0")?;

let addr = "127.0.0.1:7878";
let sent = await!(socket.send_to(THE_MERCHANT_OF_VENICE, &addr))?;
println!("Sent {} bytes to {}", sent, addr);

Important traits for RecvFrom<'socket, 'buf>
impl<'socket, 'buf> Future for RecvFrom<'socket, 'buf> type Output = Result<(usize, SocketAddr)>;

pub fn recv_from<'socket, 'buf>(
&'socket mut self,
buf: &'buf mut [u8]
) -> RecvFrom<'socket, 'buf>

Receives data from the socket.

On success, returns the number of bytes read and the origin.

Examples

#![feature(async_await, await_macro, futures_api)]
use runtime::net::UdpSocket;

let mut socket = UdpSocket::bind("127.0.0.1:0")?;

let mut buf = vec![0; 1024];
let (recv, peer) = await!(socket.recv_from(&mut buf))?;
println!("Received {} bytes from {}", recv, peer);

pub fn broadcast(&self) -> Result<bool>

Gets the value of the SO_BROADCAST option for this socket.

For more information about this option, see set_broadcast.

pub fn set_broadcast(&self, on: bool) -> Result<()>

Sets the value of the SO_BROADCAST option for this socket.

When enabled, this socket is allowed to send packets to a broadcast address.

pub fn multicast_loop_v4(&self) -> Result<bool>

Gets the value of the IP_MULTICAST_LOOP option for this socket.

For more information about this option, see set_multicast_loop_v4.

pub fn set_multicast_loop_v4(&self, on: bool) -> Result<()>

Sets the value of the IP_MULTICAST_LOOP option for this socket.

If enabled, multicast packets will be looped back to the local socket.

Note

This may not have any affect on IPv6 sockets.

pub fn multicast_ttl_v4(&self) -> Result<u32>

Gets the value of the IP_MULTICAST_TTL option for this socket.

For more information about this option, see set_multicast_ttl_v4.

pub fn set_multicast_ttl_v4(&self, ttl: u32) -> Result<()>

Sets the value of the IP_MULTICAST_TTL option for this socket.

Indicates the time-to-live value of outgoing multicast packets for this socket. The default value is 1 which means that multicast packets don't leave the local network unless explicitly requested.

Note

This may not have any affect on IPv6 sockets.

pub fn multicast_loop_v6(&self) -> Result<bool>

Gets the value of the IPV6_MULTICAST_LOOP option for this socket.

For more information about this option, see set_multicast_loop_v6.

pub fn set_multicast_loop_v6(&self, on: bool) -> Result<()>

Sets the value of the IPV6_MULTICAST_LOOP option for this socket.

Controls whether this socket sees the multicast packets it sends itself.

Note

This may not have any affect on IPv4 sockets.

pub fn ttl(&self) -> Result<u32>

Gets the value of the IP_TTL option for this socket.

For more information about this option, see set_ttl.

pub fn set_ttl(&self, ttl: u32) -> Result<()>

Sets the value for the IP_TTL option on this socket.

This value sets the time-to-live field that is used in every packet sent from this socket.

pub fn join_multicast_v4(
&self,
multiaddr: &Ipv4Addr,
interface: &Ipv4Addr
) -> Result<()>

Executes an operation of the IP_ADD_MEMBERSHIP type.

This function specifies a new multicast group for this socket to join. The address must be a valid multicast address, and interface is the address of the local interface with which the system should join the multicast group. If it's equal to INADDR_ANY then an appropriate interface is chosen by the system.

Examples

#![feature(async_await, await_macro, futures_api)]

use runtime::net::UdpSocket;
use std::net::Ipv4Addr;

let interface = Ipv4Addr::new(0, 0, 0, 0);
let mdns_addr = Ipv4Addr::new(224, 0, 0, 123);

let socket = UdpSocket::bind("127.0.0.1:0")?;
socket.join_multicast_v4(&mdns_addr, &interface)?;

pub fn join_multicast_v6(
&self,
multiaddr: &Ipv6Addr,
interface: u32
) -> Result<()>

Executes an operation of the IPV6_ADD_MEMBERSHIP type.

This function specifies a new multicast group for this socket to join. The address must be a valid multicast address, and interface is the index of the interface to join/leave (or 0 to indicate any interface).

Examples

#![feature(async_await, await_macro, futures_api)]

use runtime::net::UdpSocket;
use std::net::{Ipv6Addr, SocketAddr};

let socket_addr = SocketAddr::new(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 0).into(), 0);
let mdns_addr = Ipv6Addr::new(0xFF02, 0, 0, 0, 0, 0, 0, 0x0123) ;
let socket = UdpSocket::bind(&socket_addr)?;

socket.join_multicast_v6(&mdns_addr, 0)?;

pub fn leave_multicast_v4(
&self,
multiaddr: &Ipv4Addr,
interface: &Ipv4Addr
) -> Result<()>

Executes an operation of the IP_DROP_MEMBERSHIP type.

For more information about this option, see join_multicast_v4.

pub fn leave_multicast_v6(
&self,
multiaddr: &Ipv6Addr,
interface: u32
) -> Result<()>

Executes an operation of the IPV6_DROP_MEMBERSHIP type.

For more information about this option, see join_multicast_v6.