Trait runtime::Bytes

pub unsafe trait Bytes: Sized + Copy {
    fn uninitialized() -> Self { ... }

    fn size() -> usize { ... }

    fn as_u8_ptr(&self) -> *const u8 { ... }

    fn as_mut_u8_ptr(&mut self) -> *mut u8 { ... }
}

A trait for dealing with Bytes. Used as the underlying type for the Guarded and GuardedVec types. For a type to be able to be placed in one of these values, it must implement this trait.

Provided methods

fn uninitialized() -> Self

fn size() -> usize

fn as_u8_ptr(&self) -> *const u8

fn as_mut_u8_ptr(&mut self) -> *mut u8