Trait runtime::RedeemPallet

pub trait RedeemPallet {
    fn request_redeem<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        amount: u128, 
        btc_address: BtcAddress, 
        vault_id: &'life1 VaultId
    ) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn execute_redeem<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        redeem_id: H256, 
        merkle_proof: &'life1 [u8], 
        raw_tx: &'life2 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn cancel_redeem<'life0, 'async_trait>(
        &'life0 self, 
        redeem_id: H256, 
        reimburse: bool
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_redeem_request<'life0, 'async_trait>(
        &'life0 self, 
        redeem_id: H256
    ) -> Pin<Box<dyn Future<Output = Result<InterBtcRedeemRequest, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_vault_redeem_requests<'life0, 'async_trait>(
        &'life0 self, 
        account_id: AccountId
    ) -> Pin<Box<dyn Future<Output = Result<Vec<(H256, InterBtcRedeemRequest)>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_redeem_period<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<BlockNumber, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Required methods

fn request_redeem<'life0, 'life1, 'async_trait>(
&'life0 self,
amount: u128,
btc_address: BtcAddress,
vault_id: &'life1 VaultId
) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Request a new redeem

fn execute_redeem<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
redeem_id: H256,
merkle_proof: &'life1 [u8],
raw_tx: &'life2 [u8]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Execute a redeem request by providing a Bitcoin transaction inclusion proof

fn cancel_redeem<'life0, 'async_trait>(
&'life0 self,
redeem_id: H256,
reimburse: bool
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Cancel an ongoing redeem request

fn get_redeem_request<'life0, 'async_trait>(
&'life0 self,
redeem_id: H256
) -> Pin<Box<dyn Future<Output = Result<InterBtcRedeemRequest, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn get_vault_redeem_requests<'life0, 'async_trait>(
&'life0 self,
account_id: AccountId
) -> Pin<Box<dyn Future<Output = Result<Vec<(H256, InterBtcRedeemRequest)>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get all open redeem requests requested of the given vault

fn get_redeem_period<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<BlockNumber, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,