Crate runtime

Re-exports

pub use prometheus;

pub use sp_core;

pub use types::*;

Modules

FixedPointTraits

Primitives for the runtime modules.

cli

metadata

types

utils

Structs

AssetRegistry

BTreeSet

An ordered set based on a B-Tree.

FixedU128

A fixed point number representation in the range. Fixed Point 128 bits unsigned, range = [0.000000000000000000, 340282366920938463463.374607431768211455]

InterBtcParachain

InterBtcRuntime

LendingAssets

ShutdownReceiver

ShutdownSender

A wrapper arround a tokio broadcast channel that makes sure that listeners created after a shutdown signal has already been sent also receive the shutdown signal.

Enums

Error

RetryPolicy

SubxtError

The underlying error enum, generic over the type held by the Runtime variant. Prefer to use the Error<E> and Error aliases over using this type directly.

Constants

BLOCK_INTERVAL

BTC_RELAY_MODULE

DEFAULT_SPEC_NAME

DISABLE_DIFFICULTY_CHECK

ISSUE_MODULE

MILLISECS_PER_BLOCK

SS58_PREFIX

STABLE_BITCOIN_CONFIRMATIONS

STABLE_PARACHAIN_CONFIRMATIONS

SYSTEM_MODULE

TX_FEES

VAULT_REGISTRY_MODULE

Traits

BtcRelayPallet

CollateralBalancesPallet

CurrencyInfo

FeePallet

FixedPointNumber

Something that implements a decimal fixed point number.

IssuePallet

OraclePallet

PartialAddress

RedeemPallet

ReplacePallet

RuntimeCurrencyInfo

Fallible operations on currencies

SecurityPallet

SudoPallet

TimestampPallet

TryFromSymbol

Convert a ticker symbol into a CurrencyId at runtime

UtilFuncs

VaultRegistryPallet

Functions

notify_retry

parse_collateral_currency

parse_native_currency

parse_wrapped_currency

Type Definitions

FeeRateUpdateReceiver