Crate runtime

Modules

api

Structs

BlockTime

BlockWeight

Configuration

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

Version

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

RUNTIME_API_VERSIONS

VERSION

Type Aliases

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

Balances

BalancesConfig

FliteSystem

System

SystemConfig

Timestamp