Crate runtime

Modules

locked_memory

memories

utils

Enums

MemoryError

The different types of Error that may be encountered while using this crate

Constants

DEBUG_MSG

The memory types of this crate shall return this message when trying to debug them

Traits

Bytes

A trait for dealing with Bytes. Used as the underlying type for the Guarded and GuardedVec types. For a type to be able to be placed in one of these values, it must implement this trait.

ZeroizeOnDrop

A simple trait to force the types to call zeroize() when dropping

Derive Macros

DeriveError