Crate runtime

Modules

builtins
Builtin functions. These are usually specially generated in the parse stage, and compiled into special op codes. These operations are hard coded for the builtin types.

bytecode
Definitions for the technetium bytecode, which all programs will get compiled into to be run.

core_objects
Core objects such as integers, booleans and strings for technetium

error
Runtime errors for technetium

memory

prelude

standard
The standard library. This module contains definitions for function objects which are in scope throughout all technetium programs

Macros

func_object

func_object_void

Structs

HashableObjectRef
An object reference that's guaranteed to have a valid hash (that doesn't throw errors)

ObjectCell
The universal container object for implementers of the Object trait. Anything that implements Object should be of the form ObjectCell. ObjectCell combines interior mutability and shared ownership, so it is the primary container used for objects in technetium

ObjectRef
The main object reference type, which can be passed around to represent an object of any valid Object type

Statics

PARSED_CLARGS

Traits

LockImmutable

Object
The primary trait for objects in technetium.

OpaqueClone

RawPointer

ToAny