Struct runtime::InterBtcParachain

pub struct InterBtcParachain {
    pub native_currency_id: CurrencyId,
    pub relay_chain_currency_id: CurrencyId,
    pub wrapped_currency_id: CurrencyId,
    /* private fields */
}

Fields
native_currency_id: CurrencyIdrelay_chain_currency_id: CurrencyIdwrapped_currency_id: CurrencyId
Implementations

impl InterBtcParachain

pub async fn new<P: RpcClientT>( rpc_client: P, signer: InterBtcSigner, shutdown_tx: ShutdownSender ) -> Result<Self, Error>

pub async fn from_url( url: &str, signer: InterBtcSigner, shutdown_tx: ShutdownSender ) -> Result<Self, Error>

pub async fn from_url_with_retry( url: &str, signer: InterBtcSigner, connection_timeout: Duration, shutdown_tx: ShutdownSender ) -> Result<Self, Error>

pub async fn from_url_and_config_with_retry( url: &str, signer: InterBtcSigner, max_concurrent_requests: Option<usize>, max_notifs_per_subscription: Option<usize>, connection_timeout: Duration, shutdown_tx: ShutdownSender ) -> Result<Self, Error>

pub async fn get_finalized_block_hash(&self) -> Result<H256, Error>

pub async fn on_block<F, R>(&self, on_block: F) -> Result<(), Error>where F: Fn(InterBtcHeader) -> R, R: Future<Output = Result<(), Error>>,

Subscribe to new parachain blocks.

pub async fn wait_for_block(&self, height: u32) -> Result<(), Error>

Wait for the block at the given height Note: will always wait at least one block.

pub async fn delay_for_blocks(&self, delay: u32) -> Result<(), Error>

Sleep for delay parachain blocks

pub async fn on_event_error<E: Fn(Error)>( &self, on_error: E ) -> Result<(), Error>

Subscription service that should listen forever, only returns if the initial subscription cannot be established. Calls on_error when an error event has been received, or when an event has been received that failed to be decoded into a raw event.

Arguments

on_error - callback for decoding errors, is not allowed to take too long

pub async fn on_event<T, F, R, E>( &self, on_event: F, on_error: E ) -> Result<(), Error>where T: StaticEvent + Debug, F: FnMut(T) -> R, R: Future<Output = ()>, E: Fn(Error),

Subscription service that should listen forever, only returns if the initial subscription cannot be established. This function uses two concurrent tasks: one for the event listener, and one that calls the given callback. This allows the callback to take a long time to complete without breaking the rpc communication, which could otherwise happen. Still, since the queue of callbacks is processed sequentially, some care should be taken that the queue does not overflow. on_error is called when the event has successfully been decoded into a raw_event, but failed to decode into an event of type T

Arguments

on_event - callback for events, is allowed to sometimes take a longer time

on_error - callback for decoding error, is not allowed to take too long

pub async fn store_assets_metadata(&self) -> Result<(), Error>

pub async fn store_lend_tokens(&self) -> Result<(), Error>

pub async fn listen_for_registered_assets(&self) -> Result<(), Error>

Cache registered assets and updates

pub async fn listen_for_lending_markets(&self) -> Result<(), Error>

Cache new markets and updates

pub async fn listen_for_fee_rate_changes(&self) -> Result<(), Error>

Listen to fee_rate changes and broadcast new values on the fee_rate_update_tx channel