Struct runtime::ObjectCell

#[repr(transparent)]pub struct ObjectCell<T> where
    ObjectCell<T>: Object,  { /* fields omitted */ }

The universal container object for implementers of the Object trait. Anything that implements Object should be of the form ObjectCell. ObjectCell combines interior mutability and shared ownership, so it is the primary container used for objects in technetium

Implementations

impl<T> ObjectCell<T> where
ObjectCell<T>: Object,

pub fn new(val: T) -> Self

pub fn lock(&self)

Lock the MLRefCell inside the cell, effectively making the value immutable, giving a runtime error for any future mutation of the object. This is useful if the value will need to be used as the key in a HashMap, or in a HashSet