Trait runtime::Object

pub trait Object: Any + ToAny + OpaqueClone + RawPointer + LockImmutable {
    fn technetium_type_name(&self) -> String;

    fn technetium_clone(&self) -> RuntimeResult<ObjectRef> { ... }

    fn technetium_hash(&self) -> Option<u64> { ... }

    fn to_string(&self) -> RuntimeResult<String> { ... }

    fn get_attr(&self, _attr: String) -> RuntimeResult<ObjectRef> { ... }

    fn set_attr(&self, _attr: String, _val: ObjectRef) -> RuntimeResult<()> { ... }

    fn call_method(
        &self, 
        _method: &str, 
        _args: &[ObjectRef]
    ) -> RuntimeResult<ObjectRef> { ... }

    fn call(
        &self, 
        _args: &[ObjectRef], 
        _locals: &mut MemoryManager
    ) -> RuntimeResult<ObjectRef> { ... }

    fn make_iter(&self) -> RuntimeResult<ObjectRef> { ... }

    fn take_iter(&self) -> RuntimeResult<Option<ObjectRef>> { ... }

    fn truthy(&self) -> bool { ... }

    fn technetium_eq(&self, _other: ObjectRef) -> Option<bool> { ... }

    fn ref_eq(&self, other: ObjectRef) -> bool { ... }
}

The primary trait for objects in technetium.

Types that implement Object should be of the form ObjectCell<T> for some T. This will give all of the requirement subtraits for free.

Required methods

fn technetium_type_name(&self) -> String

A type name for an object.

Conventions are that type names are all lowercase, and use parentheses to denote "sub-types" (for example: "iterator(list)")

This function should not fail, so should return a set value.

Loading content...
Provided methods

fn technetium_clone(&self) -> RuntimeResult<ObjectRef>

Create a deep clone of an object. This is primarily used in the clone function in technetium

fn technetium_hash(&self) -> Option<u64>

Hash an object. This is not required, so the default implementation always returns None.

In implementing a hash, it's important that x == y implies that x.technetium_hash() == y.technetium_hash() to avoid logic errors.

Note that ObjectRef does not implement Hash in Rust, but HashableObjectRef does. See the docs for HashableObjectRef for more information

fn to_string(&self) -> RuntimeResult<String>

Convert an object to a String.

fn get_attr(&self, _attr: String) -> RuntimeResult<ObjectRef>

Get an attribute of an object

fn set_attr(&self, _attr: String, _val: ObjectRef) -> RuntimeResult<()>

Set an attribute of an object

fn call_method(
&self,
_method: &str,
_args: &[ObjectRef]
) -> RuntimeResult<ObjectRef>

Call a given method of an object

fn call(
&self,
_args: &[ObjectRef],
_locals: &mut MemoryManager
) -> RuntimeResult<ObjectRef>

Call a given object as a function.

This takes a memory manager primarily for the Function object, which needs to be able to reference and change locals.

fn make_iter(&self) -> RuntimeResult<ObjectRef>

Create an iterator over an object. This is used for initializing for loops.

fn take_iter(&self) -> RuntimeResult<Option<ObjectRef>>

Take from this object, assuming it is an iterator. This is used for stepping through for loops.

fn truthy(&self) -> bool

Determine whether an object is "truthy" (whether it should be treated as true when used as a boolean)

fn technetium_eq(&self, _other: ObjectRef) -> Option<bool>

Equal-as-value (like == in Python, or .equals() in Java)

fn ref_eq(&self, other: ObjectRef) -> bool

Equal-as-reference (like is in Python, or == in Java)

This is treated as a fallback in the Eq implementation for ObjectRef, primarily for sets and dictionaries

Loading content...