Trait runtime::Runtime

pub trait Runtime {
    fn curr_epoch(&self) -> ChainEpoch;

    fn randomness(&self, epoch: ChainEpoch, offset: u64) -> Randomness;

    fn immediate_caller(&self) -> &Address;

    fn validate_immediate_caller_is(&self, caller: &Address);

    fn validate_immediate_caller_accept_any(&self);

    fn validate_immediate_caller_matches(&self, caller_pattern: CallerPattern);

    fn curr_receiver(&self) -> &Address;

    fn top_level_block_winner(&self) -> &Address;

    fn acquire_state(&self) -> ActorStateHandle;

    fn success_return(&self) -> InvocOutput;

    fn value_return(&self, bytes: Vec<u8>) -> InvocOutput;

    fn abort(&self, err_exit_code: ExitCode, msg: &str);

    fn abort_arg_msg(&self, msg: String);

    fn abort_arg(&self);

    fn abort_state_msg(&self, msg: String);

    fn abort_state(&self);

    fn abort_funds_msg(&self, msg: String);

    fn abort_funds(&self);

    fn abort_api(&self, msg: String);

    fn assert(&self, cond: bool);

    fn current_balance(&self) -> TokenAmount;

    fn value_received(&self) -> TokenAmount;

    fn verify_signature(
        &self, 
        signer_actor: Address, 
        sig: Signature, 
        m: UnsignedMessage
    ) -> bool;

    fn compute(&self, id: ComputeFunctionID, args: dyn Any) -> dyn Any;

    fn send_propagating_errors(&self, input: InvocInput) -> InvocOutput;

    fn send_catching_errors(
        &self, 
        input: InvocInput
    ) -> Result<InvocOutput, ExitCode>;

    fn new_actor_address(&self) -> Address;

    fn create_actor(
        &self, 
        state_cid: &Cid, 
        a: &Address, 
        init_balance: TokenAmount, 
        constructor_params: dyn Any
    );

    fn ipld_get(&self, c: &Cid) -> Result<Vec<u8>, String>;

    fn ipld_put(&self, object: IPLDObject) -> Cid;
}

Runtime is the VM's internal runtime object. this is everything that is accessible to actors, beyond parameters.

Required methods

fn curr_epoch(&self) -> ChainEpoch

Retrieves current epoch

fn randomness(&self, epoch: ChainEpoch, offset: u64) -> Randomness

Randomness returns a (pseudo)random stream (indexed by offset) for the current epoch.

fn immediate_caller(&self) -> &Address

Not necessarily the actor in the From field of the initial on-chain Message.

fn validate_immediate_caller_is(&self, caller: &Address)

fn validate_immediate_caller_accept_any(&self)

fn validate_immediate_caller_matches(&self, caller_pattern: CallerPattern)

fn curr_receiver(&self) -> &Address

The address of the actor receiving the message.

fn top_level_block_winner(&self) -> &Address

The actor who mined the block in which the initial on-chain message appears.

fn acquire_state(&self) -> ActorStateHandle

fn success_return(&self) -> InvocOutput

Return successfully from invocation.

fn value_return(&self, bytes: Vec<u8>) -> InvocOutput

Return from invocation with a value.

fn abort(&self, err_exit_code: ExitCode, msg: &str)

Throw an error indicating a failure condition has occurred, from which the given actor code is unable to recover.

fn abort_arg_msg(&self, msg: String)

Calls Abort with InvalidArguments error.

fn abort_arg(&self)

fn abort_state_msg(&self, msg: String)

Calls Abort with InconsistentState error.

fn abort_state(&self)

fn abort_funds_msg(&self, msg: String)

Calls Abort with InsufficientFunds error.

fn abort_funds(&self)

fn abort_api(&self, msg: String)

Calls Abort with RuntimeAPIError. For internal use only (not in actor code).

fn assert(&self, cond: bool)

Check that the given condition is true (and call Abort if not).

fn current_balance(&self) -> TokenAmount

Retrieves current balance in VM.

fn value_received(&self) -> TokenAmount

Retrieves value received in VM.

fn verify_signature(
&self,
signer_actor: Address,
sig: Signature,
m: UnsignedMessage
) -> bool

fn compute(&self, id: ComputeFunctionID, args: dyn Any) -> dyn Any

Run a (pure function) computation, consuming the gas cost associated with that function. This mechanism is intended to capture the notion of an ABI between the VM and native functions, and should be used for any function whose computation is expensive.

fn send_propagating_errors(&self, input: InvocInput) -> InvocOutput

Send allows the current execution context to invoke methods on other actors in the system.

fn send_catching_errors(
&self,
input: InvocInput
) -> Result<InvocOutput, ExitCode>

fn new_actor_address(&self) -> Address

Computes an address for a new actor. The returned address is intended to uniquely refer to the actor even in the event of a chain re-org (whereas an ID-address might refer to a different actor after messages are re-ordered).

fn create_actor(
&self,
state_cid: &Cid,
a: &Address,
init_balance: TokenAmount,
constructor_params: dyn Any
)

Create an actor in the state tree. May only be called by InitActor.

fn ipld_get(&self, c: &Cid) -> Result<Vec<u8>, String>

fn ipld_put(&self, object: IPLDObject) -> Cid

Loading content...