Trait runtime::Signer

pub trait Signer<T, E> where
    T: Config,
    E: SignedExtra<T>, {
    fn account_id(&self) -> &<T as Config>::AccountId;

    fn nonce(&self) -> Option<<T as Config>::Index>;

    fn sign<'life0, 'async_trait>(
        &'life0 self, 
        extrinsic: SignedPayload<Encoded, <E as SignedExtra<T>>::Extra>
    ) -> Pin<Box<dyn Future<Output = Result<UncheckedExtrinsic<<T as Config>::Address, Encoded, <T as Config>::Signature, <E as SignedExtra<T>>::Extra>, String>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Extrinsic signer.

Required methods

fn account_id(&self) -> &<T as Config>::AccountId

Returns the account id.

fn nonce(&self) -> Option<<T as Config>::Index>

Optionally returns a nonce.

fn sign<'life0, 'async_trait>(
&'life0 self,
extrinsic: SignedPayload<Encoded, <E as SignedExtra<T>>::Extra>
) -> Pin<Box<dyn Future<Output = Result<UncheckedExtrinsic<<T as Config>::Address, Encoded, <T as Config>::Signature, <E as SignedExtra<T>>::Extra>, String>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Takes an unsigned extrinsic and returns a signed extrinsic.

Some signers may fail, for instance because the hardware on which the keys are located has refused the operation.